refactor: 회원가입 후크 분리 (#148)
